added optional output_bursts, and capped the size of exponential distribution samples
Also changed nu to reflect the horizontal accuracy directly with the name ha, and changed the column 'ha' to indicate the horizontal accuracy, which should be the parameter given (a priori) and not a realized error (a posteriori).

Perhaps some robustness experiment could use the info on the realized error, but then they can check the complete noiseless trajectory.