Implemented safe filter on read for pyarrow with 2 tests
Now users can specify a column to filter on upon reading files, which is useful to read chunks of a partitioned dataset. Implemented two unit tests for it.

The filters can run into data type issues and the user might pass a traj_cols key instead of an actual column name. Both cases are handled.