Robustness experiments (#125)
The robustness research notebook (`robustness-of-algos-paper.ipynb`) is
refactored to work with the most up-to-date `nomad` code in the main
branch. Minor bug corrections are made in the nomad code base to correct
issues that arose when running the robustness code.

As Thomas is transitioning out of this project, the code to run the
robustness experiments also now contains better documentation and code
comments to guide researchers in running the experiments.

Currently, there are three experiments coded up:
- Exp. 1A demonstrates merging in DBScan and Lachesis over different
sparsity patterns.
- Exp. 1B demonstrates splitting and missing in DBScan and Lachesis over
different sparsity patterns.
- Exp. 2 compares the F1 score of different algorithms over different
completeness and noise patterns in realistic (3-week) datasets.

The code can be run in a parallelized setting (e.g. AWS Sagemaker) to
reproduce the simulation results. Alternatively, .pkl files contain the
most recent runs of the simulations and can be accessed to reproduce
charts.

IN ADDITION, this pull includes code for generating a rasterized "garden
city" version of a real city (e.g., Philadelphia) through the use of
street arterials and Open Street Maps. The `RealCityGenerator` class
defined in `virtual_philly.ipynb` provides the methods necessary for
this application. An example usage is also provided in
`virtual_philly.ipynb`. If all checks out, the `RealCityGenerator` class
should be moved into `city_gen.py`.

---------

Co-authored-by: Thomas Li <thomli@sas.upenn.edu>
Co-authored-by: Jorge Francisco Barreras <franciscojosebarreras@gmail.com>
Co-authored-by: Jorge Barreras <fbarrer@sas.upenn.edu>
Co-authored-by: Jorge Francisco Barreras <30353961+paco-barreras@users.noreply.github.com>